Fix random_period() returning negative values causing iterator stalls
random_period() used rand::random::<isize>() which is negative ~50% of
the time. Rust's % operator preserves the sign of the dividend, so the
result was negative half the time. This caused the while loop in
record_read_sample() to add negative values to byte_count, turning it
into a symmetric random walk that could spin for seconds or minutes
before crossing zero.

The original C++ LevelDB uses Random::Uniform() which returns [0, n).
Use rand::random::<usize>() to match that behavior.

+++ src/db_iter.rs
@@ -299,7 +299,7 @@ impl LdbIterator for DBIterator {
 }
 
 fn random_period() -> isize {
-    rand::random::<isize>() % (2 * READ_BYTES_PERIOD)
+    (rand::random::<usize>() % (2 * READ_BYTES_PERIOD as usize)) as isize
 }
 
 #[cfg(test)]
